http1: use size_t for the length argument of Parser::execute (#47265)
Commit Message: http1: use size_t for the length argument of
Parser::execute

Additional Description:

`Http::Http1::Parser::execute()` declares its length argument as `int`,
but every caller passes a `size_t`:

| Call site | Argument | Type |
| --- | --- | --- |
| `ConnectionImpl::dispatchSlice()`
(`source/common/http/http1/codec_impl.cc:713`) | `RawSlice::len_` |
`size_t` |
| generic_proxy HTTP/1 codec
(`source/extensions/filters/network/generic_proxy/codecs/http1/config.cc:278`)
| `RawSlice::len_` | `size_t` |
| http_11_proxy transport socket
(`source/extensions/transport_sockets/http_11_proxy/connect.cc:25`) |
`response_payload.length()` | `size_t` |
| http_inspector listener filter
(`source/extensions/filters/listener/http_inspector/http_inspector.cc:138,158`)
| `new_data.length()` | `size_t` |

Each of these narrows implicitly (`-Wshorten-64-to-32`). A length above
`INT_MAX` truncates to a negative `int`; every guard in
`BalsaParser::execute()` tests `> 0` or `== 0`
(`balsa_parser.cc:214,235,242`), so a negative value passes all of them
and is then handed to `BalsaFrame::ProcessInput()`, where it widens back
into a very large `size_t`. `LegacyHttpParserImpl` has the same shape.

This is latent rather than triggerable. Read slices come from
`reserveForRead()` and never approach 2 GB, so the parser does not see
such a length today. But that is an allocation constant rather than an
enforced invariant, and the bound on the returned count in
`codec_impl.cc` is a debug-only `ASSERT`, so the case for fixing is the
wrong signature rather than a live bug.

Both underlying parsers already take `size_t`:

* `size_t BalsaFrame::ProcessInput(const char* input, size_t size)`
* `size_t http_parser_execute(http_parser*, const http_parser_settings*,
const char* data, size_t len)`

so `int` was purely an artifact of Envoy's own `Parser` interface,
sitting between two `size_t` APIs. Widening it to `size_t` removes the
narrowing at all four call sites without changing behaviour, and no call
site needs to change. The `> 0` and `== 0` guards in
`BalsaParser::execute()` remain correct for an unsigned length, and
`http_inspector`'s `ssize_t rc = parser_->execute(...)` is unaffected
since the return type was always `size_t`.

`BalsaParser` and `LegacyHttpParserImpl` are the only implementations of
`Parser`; both are updated.

Risk Level: Low — signature-only change; no behavioural change on any
reachable path, and no call site required modification.

Testing: Existing coverage exercises all of this.
`test/common/http/http1/balsa_parser_test.cc`,
`balsa_parser_benchmark_test.cc` and
`test/extensions/transport_sockets/http_11_proxy/connect_test.cc`
already pass `size_t` values (`request.size()`, `headers.length()`) into
`execute()`, so they were themselves relying on the narrowing conversion
and now type-check exactly. No new test is added because there is no
reachable behaviour change to assert: producing a >2 GB slice is not
possible through Envoy's read path. The parallel fix for the other half
of this issue (f08dad9cef) landed on the same basis.

Docs Changes: N/A

Release Notes: N/A — not user-facing.

Platform Specific Features: N/A

Fixes #46505

The `isUrlValid()` half of that issue was fixed separately in
f08dad9cef; this is the remaining `size_t`-to-`int` narrowing item.

---

Generative AI disclosure, per the
[policy](https://github.com/envoyproxy/envoy/blob/main/CONTRIBUTING.md?plain=1#L41):
this change was prepared with AI assistance (Claude Code). I have
reviewed and understand the change, and take full ownership of it. It is
a five-line signature change; the call-site inventory and the upstream
`ProcessInput`/`http_parser_execute` signatures above were each verified
directly against the source rather than assumed.

+++ source/common/http/http1/balsa_parser.h
@@ -24,7 +24,7 @@ public:
   ~BalsaParser() override = default;
 
   // Http1::Parser implementation
-  size_t execute(const char* slice, int len) override;
+  size_t execute(const char* slice, size_t len) override;
   void resume() override;
   CallbackResult pause() override;
   ParserStatus getStatus() const override;
